fix: move interface to root folder
