+++ TmsRunner/App.cs
@@ -4,7 +4,6 @@ using TmsRunner.Entities;
 using TmsRunner.Entities.Configuration;
 using TmsRunner.Managers;
 using TmsRunner.Services;
-using TmsRunner.Services.Interfaces;
 
 namespace TmsRunner;
