+++ TmsRunner/Program.cs
@@ -21,7 +21,6 @@ using TmsRunner.Handlers;
 using TmsRunner.Managers;
 using TmsRunner.Services;
 using TmsRunner.Services.Implementations;
-using TmsRunner.Services.Interfaces;
 using TmsRunner.Utils;
 using ConfigurationManager = TmsRunner.Managers.ConfigurationManager;
