+++ TmsRunner/Services/Implementations/TestRunContextService.cs
@@ -1,5 +1,4 @@
 using TestIT.ApiClient.Model;
-using TmsRunner.Services.Interfaces;
 
 namespace TmsRunner.Services.Implementations;
